Use MHD_USE_PIPE_FOR_SHUTDOWN when listen->type == LOGFANOUTD_LISTEN_FD
When exiting listening socket is used, shutdown() call on it has to be avoided.
The following issue with systemd happens:
"Got POLLHUP on a listening socket. The service probably invoked shutdown() on it, and should better not do that."

libmicrohttpd docs say that MHD_USE_PIPE_FOR_SHUTDOWN option guarantees that MHD will not call shutdown() on the listen socket.

Close #9